Check for invalid jrebel configs

+++ lrbuild.sh
@@ -8,6 +8,14 @@ if [ ! -f pom.xml ]; then
 	exit 1
 fi
 
+BAD_REBEL=`find . -name rebel.xml -exec grep -lL "$WORKSPACE_LOCATION" {} \;`
+
+if [ -n "$BAD_REBEL" ]; then
+	echo "We have bad rebel files"
+	echo ${BAD_REBEL}
+	exit 1
+fi
+
 find . -type d -name .sass_cache -exec rm -r {} \;
 
 mvn clean || exit 1
